fixed file deletion

+++ routes/file.js
@@ -80,10 +80,17 @@ module.exports = (db, bucket) => {
 
     router.delete('/delete/:id', async(req, res) => {
         try {
-            const idField = new ObjectId(req.params.id);            
-            const result = await files_collection.findOneAndDelete({_id : idField});            
-            await chunks_collection.findOneAndDelete({files_id: idField});
-            res.send("file deleted");
+            const idField = new ObjectId(req.params.id);
+            const file = await files_collection.find({_id : idField});
+            if(file){
+                const result = await files_collection.findOneAndDelete({_id : idField});            
+                await chunks_collection.findOneAndDelete({files_id: idField});
+                res.send("file deleted");
+            }
+            else{
+                res.send("file not found");
+            }   
+            
         } catch (error) {
             res.send("Error deleting file");
         }
